Renamed Docs.json to meta and added puml support
